vyny/requirements.txt
asgiref==3.8.1
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloud-init==24.3.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
dateparser==1.2.0
dbus-python==1.2.18
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.1
djangorestframework==3.15.2
filelock==3.16.0
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.10.0
greenlet==3.0.3
httplib2==0.20.2
hyperlink==21.0.0
idna==3.7
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
matplotlib==3.9.1
more-itertools==8.10.0
mpmath==1.3.0
netifaces==0.11.0
networkx==3.4.2
numpy==2.0.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.3.2
psutil==6.1.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2024.1
PyYAML==5.4.1
regex==2024.5.15
requests==2.32.3
scipy==1.14.1
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.1
sympy==1.13.1
systemd-python==234
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
triton==3.1.0
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ubuntu-pro-client==8001
ufw==0.36.1
ultralytics==8.3.20
ultralytics-thop==2.0.9
unattended-upgrades==0.1
urllib3==2.2.2
virtualenv==20.26.4
wadllib==1.3.6
Werkzeug==3.0.3
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
